Time for another huge update.
- Most of the network code is completed (I hope). It relies on a multithreaded programming approach so that files can be transferred in the background.  This has not been fully tested and is likely to still be broken. Though in theory it should work!

- factored a lot of the code and the so that we can (potentially) have multiple network games hosted on the same machine. This has led to a lot of "complicated" pointers all over the place, but they will all disappear in C++.

- renamed the enet/time.h to enet/_time.h since the naming of this file was interfering with the loading of the C header <time.h>.  This probably explains Zefz problem with using the time(NULL) function.

The game should run, you should be able to check out the network menu section and contact another computer (though no network play, yet).  There is a problem with the particles not displaying and not interacting with the mesh, as well as some wack script behavior.  I'm 100% certain that all of that is because I have not placed the appropriate pointers in all the function calls. #$%^ C for not detecting these mismatches!!!