Added some code to minimize the chance of unexpected NULL pointers in Menu.c. This is the reason that several of the menus, including the Network menu did not function properly.
There may be some problem reading in the spawn info for Bishopia City (the only place I have seen the error). The game is getting some bad data when trying to spawn the guard captain. This was causing the program to crash, but now it just ignores the bad input.