Fixed some problems with spawning into multiplayer modules. You can spawn into Bishopia City with at least one character.
The "outdoor lighting" has been modified so that it will only appear in modules where wawalite.txt has an expansion "[MOON] 1". As none of the modules have this now, they all the meshes appear dark. The dynamic lighting is not lighting the mesh at the moment, so it will be almost completely dark...