Was able to build and run the project using Code::Blocks. The IDE files have been updated.
Most of the gcc warnings have been eliminated. It still warns about const being applied to return types of functions, and about some static functions not being declared (from the definition of INLINE as "static inline" in Linux).  The const on the return types will be kept since it will be inportant for C++.